Push mouth_as_structures code into pesca_base, fire off salt run
