+++ model/catalog.txt
@@ -444,14 +444,8 @@ run_salt_20160520-v115 (farm)
 run_salt_20160520-v116 (farm)
   Two structure, increased friction. Thought I dropped a structure but
   apparently not. Results were good, though.
-  
 
+run_salt_20160520-v117 (farm)
+  100 structures. 2D was promising...
 
-
-
-  
-
-
-
-
